feat: 车辆管理

+++ service/src/main/java/cn/com/goodlan/its/service/system/college/CollegeServiceImpl.java
@@ -56,6 +56,9 @@ public class CollegeServiceImpl implements CollegeService {
             if (college.get().hasUser()) {
                 throw new BusinessException("此学院存在用户,不允许删除");
             }
+            if (college.get().hasVehicle()) {
+                throw new BusinessException("此学院存在车辆,不允许删除");
+            }
             collegeRepository.deleteById(id);
         }
     }
